Make stability `sgm` an installed dependency
Previously we had a copy of the Stability `generative_models` repo in
`src/mindeye_utils`. Instead, add the `sgm` package as a dependency,
with pinned sha to the Nov 2023 SD-Turbo release (which is when we
copied it, based on inspecting the code). Update `sgm` package imports
in the downstream notebooks.